separate flapping management into a class
